refactor(stella-store): split DDL and migrations out of lib.rs
Pure mechanical split of the 4.6k-line lib.rs, public API byte-identical:

- src/ddl.rs — every table/index DDL at the current schema shape (the
  consts plus the three name-parameterized DDL builders).
- src/migrations.rs — the schema-versioning machinery: MIGRATIONS,
  SCHEMA_VERSION, create_latest_schema, the per-version migrate fns, and
  the transactional runner apply_migration.

Everything moved is crate-private, re-imported at the crate root, so no
stella_store:: path any other crate consumes changes. Zero test edits —
the full store suite (including every per-version legacy-upgrade test)
passes untouched. lib.rs gains a module-map comment for the new layout.
