+++ stella-store/src/lib.rs
@@ -82,11 +82,28 @@ use rusqlite::{Connection, OptionalExtension, params};
 use serde::Serialize;
 use stella_protocol::{AgentEvent, TaskItem, TaskStatus, ToolOutput};
 
+// Module map — this file holds the row types, the `Store` handle and its
+// query surface, and the tests; everything else is split by concern:
+//   ddl        (crate-private) every table/index DDL at the CURRENT schema
+//   migrations (crate-private) versioned upgrades + the fresh-file bootstrap
+//   catalog    `catalog.db` — user-tier model catalog (slugs, pricing)
+//   journal    append-only per-session sidecar journal (crash-safe resume)
+//   notify     persist-until-read cross-session notifications
+//   sessions   cross-process session registry (one JSON file per session)
+//   usage      `usage.db` — user-tier cross-project telemetry aggregate
+mod ddl;
+mod migrations;
+
 pub mod catalog;
 pub mod journal;
 pub mod notify;
 pub mod sessions;
 pub mod usage;
+
+use ddl::TABLES;
+use migrations::{
+    MIGRATIONS, SCHEMA_VERSION, any_store_table_exists, apply_migration, create_latest_schema,
+};
 
 pub use catalog::CatalogStore;
 // The sidecar journal's writer is deliberately NOT re-exported at the top
@@ -503,701 +520,6 @@ fn harden_workspace_dir(dir: &Path, created: bool) -> Result<()> {
         let _ = file.write_all(b"*.db\n*.db-wal\n*.db-shm\nreflections.jsonl\n");
     }
     Ok(())
-}
-
-/// One schema migration: upgrades an existing database exactly one
-/// `user_version` step, inside the transaction the runner opened for it.
-/// The runner stamps the new version and commits; the migration only
-/// reshapes.
-type Migration = fn(&rusqlite::Transaction<'_>) -> Result<()>;
-
-/// Ordered migration list for EXISTING databases: `MIGRATIONS[i]` upgrades
-/// a file at `user_version` i to i + 1. Fresh files never run these — they
-/// get [`create_latest_schema`] and are stamped at [`SCHEMA_VERSION`]
-/// directly.
-const MIGRATIONS: [Migration; 8] = [
-    // v0 → v1: dedupe events/telemetry, then retrofit the UNIQUE keys
-    // their write paths have always assumed.
-    migrate_v0_to_v1,
-    // v1 → v2: files_touched grows line-delta totals + the JSON audit log,
-    // and the UNIQUE (execution_id, path) key.
-    migrate_v1_to_v2,
-    // v2 → v3: the memory_citations table and the `rules` table
-    // (extension-authored workspace rules for the stella-core rules
-    // engine) — both purely additive; no existing table changes shape.
-    migrate_v2_to_v3,
-    // v3 → v4: the agent_uses invocation log (purely additive).
-    migrate_v3_to_v4,
-    // v4 → v5: the skill_usage invocation log (purely additive — SKILLS tab).
-    migrate_v4_to_v5,
-    // v5 → v6: the additive `mcp_usage` table (per-call MCP tool telemetry).
-    // Renumbered from v4 at cascade-merge behind the agent_uses (v4) and
-    // skill_usage (v5) migrations; SCHEMA_VERSION follows MIGRATIONS.len().
-    migrate_v5_to_v6,
-    // v6 → v7: the data-plane tables (all purely additive) — `tool_calls`
-    // (normalized per-call log), `execution_reflection` (per-turn self-review
-    // tied to the prompt), and `reflections` (unified durable lessons).
-    migrate_v6_to_v7,
-    // v7 → v8: the session plane — `executions` grows the nullable
-    // `session_id` link (+ its by-session index), plus the additive `tasks`
-    // (per-session task-board snapshot) and `pull_requests` tables.
-    migrate_v7_to_v8,
-];
-
-/// The schema version this build writes — the `PRAGMA user_version` of
-/// every database it has opened. Version 0 is the legacy pre-versioning
-/// shape (and the default stamp of a fresh empty file, which is why fresh
-/// detection also probes for tables).
-const SCHEMA_VERSION: i64 = MIGRATIONS.len() as i64;
-
-/// Every table the store owns — the allowlist for [`Store::count`] and the
-/// fresh-file probe in [`Store::migrate`].
-const TABLES: [&str; 17] = [
-    "executions",
-    "events",
-    "telemetry",
-    "files_touched",
-    "memory_citations",
-    "rules",
-    "mcp_usage",
-    "file_locks",
-    "graph_nodes",
-    "graph_edges",
-    "agent_uses",
-    "skill_usage",
-    "tool_calls",
-    "execution_reflection",
-    "reflections",
-    "tasks",
-    "pull_requests",
-];
-
-/// `executions` DDL at [`SCHEMA_VERSION`] — the spine every other table
-/// keys off, one row per run/goal/turn. `session_id` (v8) is the nullable
-/// cross-process session registry id ([`SessionRecord::id`]) stamped by
-/// [`Store::set_execution_session`] right after the row is opened, linking
-/// per-turn executions back to their session so
-/// [`Store::session_events`] can reassemble the full journal; NULL for rows
-/// persisted before v8 or for runs outside a registered session. The
-/// by-session index is that reader's access path (filter on session_id,
-/// scan in id order). `IF NOT EXISTS` on both so the batch also tolerates a
-/// partial file that already grew them.
-const EXECUTIONS_DDL: &str = "CREATE TABLE IF NOT EXISTS executions (
-       id INTEGER PRIMARY KEY AUTOINCREMENT,
-       kind TEXT NOT NULL,
-       prompt TEXT NOT NULL,
-       provider TEXT NOT NULL,
-       model TEXT NOT NULL,
-       started_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP,
-       finished_at TEXT,
-       outcome TEXT,
-       cost_usd REAL NOT NULL DEFAULT 0,
-       session_id TEXT
-     );
-     CREATE INDEX IF NOT EXISTS executions_by_session
-       ON executions(session_id, id);";
-
-/// Tables whose shape has not changed since v0. `IF NOT EXISTS` keeps one
-/// batch usable both for fresh files and for filling gaps in partial legacy
-/// files (a v0 file only holds what its era's code created).
-const UNCHANGED_TABLES: &str = "CREATE TABLE IF NOT EXISTS file_locks (
-       path TEXT PRIMARY KEY,
-       holder TEXT NOT NULL,
-       acquired_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-     );
-     CREATE TABLE IF NOT EXISTS graph_nodes (
-       id TEXT PRIMARY KEY,
-       label TEXT NOT NULL,
-       properties TEXT NOT NULL DEFAULT '{}'
-     );
-     CREATE TABLE IF NOT EXISTS graph_edges (
-       src TEXT NOT NULL,
-       dst TEXT NOT NULL,
-       edge_type TEXT NOT NULL,
-       properties TEXT NOT NULL DEFAULT '{}'
-     );";
-
-/// `events` DDL at [`SCHEMA_VERSION`], parameterized over the table name
-/// because the v0 → v1 rebuild first creates it under a scratch name.
-///
-/// UNIQUE (execution_id, seq): one row per position in an execution's event
-/// stream. The drain loop owns a monotonically increasing `seq` per
-/// execution and replay reads `(execution_id, seq)` back in order, so a
-/// duplicate position is a double-write, not data — the constraint turns it
-/// into an error instead of a silently corrupted replay. Its implicit index
-/// is exactly the replay access path (superseding the pre-v1 non-unique
-/// `events_by_execution` index, which is why no separate index exists).
-fn events_ddl(table: &str) -> String {
-    format!(
-        "CREATE TABLE {table} (
-           execution_id INTEGER NOT NULL,
-           seq INTEGER NOT NULL,
-           ts TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP,
-           event_type TEXT NOT NULL,
-           payload TEXT NOT NULL,
-           UNIQUE (execution_id, seq)
-         );"
-    )
-}
-
-/// `files_touched` DDL at [`SCHEMA_VERSION`] — see [`events_ddl`] for why it
-/// is name-parameterized.
-///
-/// UNIQUE (execution_id, path): one session record per normalized path —
-/// the ledger aggregates every touch of a file into one record before
-/// persisting, so a duplicate path is a double-write, not data. `events` is
-/// the ordered JSON audit log (`[{event, reason, lines_added,
-/// lines_removed}, …]`); rows persisted before v2 carry the backfill
-/// defaults (zero deltas, empty log).
-fn files_touched_ddl(table: &str) -> String {
-    format!(
-        "CREATE TABLE {table} (
-           execution_id INTEGER NOT NULL,
-           path TEXT NOT NULL,
-           ops TEXT NOT NULL,
-           lines_added INTEGER NOT NULL DEFAULT 0,
-           lines_removed INTEGER NOT NULL DEFAULT 0,
-           events TEXT NOT NULL DEFAULT '[]',
-           UNIQUE (execution_id, path)
-         );"
-    )
-}
-
-/// `telemetry` DDL at [`SCHEMA_VERSION`] — see [`events_ddl`] for why it is
-/// name-parameterized.
-///
-/// UNIQUE (execution_id, step): one row per committed model call —
-/// `StepUsage` is emitted exactly once per step that lands. `drift_samples`
-/// treats `(execution_id, step)` as insertion order and `usage_stats` sums
-/// tokens/cost per execution, so a duplicate step double-counts money.
-fn telemetry_ddl(table: &str) -> String {
-    format!(
-        "CREATE TABLE {table} (
-           execution_id INTEGER NOT NULL,
-           step INTEGER NOT NULL,
-           ts TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP,
-           provider TEXT NOT NULL,
-           model TEXT NOT NULL,
-           input_tokens INTEGER NOT NULL,
-           estimated_input_tokens INTEGER NOT NULL DEFAULT 0,
-           output_tokens INTEGER NOT NULL,
-           cache_read_tokens INTEGER NOT NULL,
-           cache_miss_tokens INTEGER NOT NULL,
-           cache_write_tokens INTEGER NOT NULL DEFAULT 0,
-           cost_usd REAL NOT NULL,
-           duration_ms INTEGER NOT NULL,
-           retries INTEGER NOT NULL,
-           tool_calls INTEGER NOT NULL,
-           UNIQUE (execution_id, step)
-         );"
-    )
-}
-
-/// `rules` DDL — one row per extension-authored workspace rule, keyed by
-/// rule id (the analog of a rule file's filename stem). `contents` is the
-/// FULL rule markdown in the `.stella/rules/*.md` authoring format
-/// (optional `---` frontmatter — `description:`/`guard-*:` keys — plus the
-/// rule statement body); the store never parses it, `stella_core::rules`
-/// does. `source` is an opaque label naming the writer (extension/provider
-/// id). `IF NOT EXISTS` so one batch serves both the fresh-file schema and
-/// the v2 → v3 migration.
-const RULES_TABLE: &str = "CREATE TABLE IF NOT EXISTS rules (
-       rule_id TEXT PRIMARY KEY,
-       contents TEXT NOT NULL,
-       source TEXT NOT NULL,
-       created_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP,
-       updated_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-     );";
-
-/// `drift_samples` filters (provider, model) and sorts (execution_id DESC,
-/// step DESC) at EVERY session start, over a table that grows one row per
-/// model call forever — without this index it full-scans. Non-unique on
-/// purpose: uniqueness lives on the (execution_id, step) key; this is the
-/// query's covering access path.
-const TELEMETRY_INDEX: &str = "CREATE INDEX IF NOT EXISTS telemetry_by_model
-       ON telemetry(provider, model, execution_id, step);";
-
-/// `memory_citations` DDL at [`SCHEMA_VERSION`].
-///
-/// UNIQUE (execution_id, memory_id): one citation per memory per execution —
-/// the session ledger keeps only the model's latest judgment of a memory
-/// before persisting, so a duplicate pair is a double-write, not data.
-/// `truthful` is 0/1. The by-memory index is the access path of
-/// [`Store::memory_citation_stats`], which scans per memory in citation
-/// order; the UNIQUE key's implicit (execution_id, …) index can't serve it.
-const MEMORY_CITATIONS_DDL: &str = "CREATE TABLE IF NOT EXISTS memory_citations (
-       execution_id INTEGER NOT NULL,
-       memory_id TEXT NOT NULL,
-       ts TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP,
-       useful_score INTEGER NOT NULL,
-       truthful INTEGER NOT NULL,
-       remark TEXT NOT NULL DEFAULT '',
-       UNIQUE (execution_id, memory_id)
-     );
-     CREATE INDEX IF NOT EXISTS memory_citations_by_memory
-       ON memory_citations(memory_id, execution_id);";
-
-/// `agent_uses` DDL at [`SCHEMA_VERSION`] — the agent-invocation log
-/// ([`AgentUseRow`]): one row per invocation of an installed agent
-/// definition, attributed to the execution it ran under and to the
-/// definition's pinned `version` at invocation time. Deliberately **not**
-/// UNIQUE on any key: invoking the same agent-version twice in one execution
-/// is two real events, and the drain-per-execution write path never
-/// double-writes a drained event. `IF NOT EXISTS` keeps the one DDL usable
-/// for both the fresh-file path and the additive v3 → v4 migration.
-const AGENT_USES_DDL: &str = "CREATE TABLE IF NOT EXISTS agent_uses (
-       execution_id INTEGER NOT NULL,
-       agent TEXT NOT NULL,
-       version INTEGER NOT NULL,
-       reason TEXT NOT NULL DEFAULT '',
-       ts TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-     );
-     CREATE INDEX IF NOT EXISTS agent_uses_by_agent
-       ON agent_uses(agent, version, execution_id);";
-
-/// `skill_usage` DDL at [`SCHEMA_VERSION`] — per-execution skill-version
-/// invocation telemetry (SKILLS tab), the exact analogue of [`AGENT_USES_DDL`].
-/// Append-only: one row per skill applied in a turn, no UNIQUE key. The
-/// by-skill index serves per-skill/version aggregate queries.
-const SKILL_USAGE_DDL: &str = "CREATE TABLE IF NOT EXISTS skill_usage (
-       execution_id INTEGER NOT NULL,
-       skill TEXT NOT NULL,
-       version INTEGER NOT NULL,
-       reason TEXT NOT NULL DEFAULT '',
-       ts TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-     );
-     CREATE INDEX IF NOT EXISTS skill_usage_by_skill
-       ON skill_usage(skill, version, execution_id);";
-
-/// `mcp_usage` DDL at [`SCHEMA_VERSION`].
-///
-/// A per-call log (NOT a per-key aggregate like `files_touched`): the same
-/// server+tool called twice is two rows. UNIQUE (execution_id, seq) is the
-/// house double-write guard (the `events` pattern) — `seq` is the row's index
-/// in an execution's drained batch, so re-persisting the same drained batch is
-/// an error, not a silent double-count. `called_at_ms` is the call time
-/// captured at the tool call (not the drain time). The by-server index is the
-/// access path of [`Store::mcp_usage_stats`].
-const MCP_USAGE_DDL: &str = "CREATE TABLE IF NOT EXISTS mcp_usage (
-       execution_id INTEGER NOT NULL,
-       seq INTEGER NOT NULL,
-       server TEXT NOT NULL,
-       tool TEXT NOT NULL,
-       reason TEXT NOT NULL DEFAULT '',
-       called_at_ms INTEGER NOT NULL,
-       UNIQUE (execution_id, seq)
-     );
-     CREATE INDEX IF NOT EXISTS mcp_usage_by_server
-       ON mcp_usage(server, tool);";
-
-/// `tool_calls` DDL at [`SCHEMA_VERSION`] — one queryable row per tool call
-/// (native, MCP, skill, or agent), normalized from the append-only `events`
-/// stream (`tool_start` + `tool_result`) so the dashboard can query call
-/// histograms without JSON-scanning the event log. Large outputs are NOT
-/// stored here — only shape, timing, and success (`bytes_out` records the
-/// result size, not the result). UNIQUE (execution_id, seq) is the house
-/// double-write guard. The by-name index is the access path for usage
-/// histograms (e.g. "grep called N times, graph_query zero").
-const TOOL_CALLS_DDL: &str = "CREATE TABLE IF NOT EXISTS tool_calls (
-       execution_id INTEGER NOT NULL,
-       seq INTEGER NOT NULL,
-       call_id TEXT NOT NULL DEFAULT '',
-       name TEXT NOT NULL,
-       surface TEXT NOT NULL DEFAULT 'native',
-       args_json TEXT NOT NULL DEFAULT '{}',
-       args_digest TEXT NOT NULL DEFAULT '',
-       reason TEXT NOT NULL DEFAULT '',
-       ok INTEGER NOT NULL DEFAULT 1,
-       error TEXT NOT NULL DEFAULT '',
-       bytes_out INTEGER NOT NULL DEFAULT 0,
-       duration_ms INTEGER NOT NULL DEFAULT 0,
-       ts TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP,
-       UNIQUE (execution_id, seq)
-     );
-     CREATE INDEX IF NOT EXISTS tool_calls_by_name
-       ON tool_calls(name, execution_id);";
-
-/// `execution_reflection` DDL at [`SCHEMA_VERSION`] — the agent's own
-/// assessment of ONE turn, tied 1:1 to its execution (and thus to
-/// `executions.prompt`). Pairs the model's self-view (`delivered`,
-/// `self_rating`, `what_went_well`, `what_to_improve`, `critique`) with the
-/// objective companions (`produced_output`, `wrote_files`, `truncated`) so a
-/// self-silent, zero-output turn is visibly a failure even if the model would
-/// rate itself kindly.
-const EXECUTION_REFLECTION_DDL: &str = "CREATE TABLE IF NOT EXISTS execution_reflection (
-       execution_id INTEGER PRIMARY KEY,
-       prompt TEXT NOT NULL DEFAULT '',
-       delivered INTEGER,
-       self_rating INTEGER,
-       what_went_well TEXT NOT NULL DEFAULT '',
-       what_to_improve TEXT NOT NULL DEFAULT '',
-       critique TEXT NOT NULL DEFAULT '',
-       produced_output INTEGER NOT NULL DEFAULT 0,
-       wrote_files INTEGER NOT NULL DEFAULT 0,
-       truncated INTEGER NOT NULL DEFAULT 0,
-       recorded_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-     );";
-
-/// `reflections` DDL at [`SCHEMA_VERSION`] — the durable, unified home for
-/// lessons and self-critiques (superset of the loose `.stella/reflections.jsonl`
-/// and the context.db memory nodes). `execution_id` is NULL for cross-turn
-/// lessons; `domains` is a JSON array of domain tags.
-const REFLECTIONS_DDL: &str = "CREATE TABLE IF NOT EXISTS reflections (
-       id INTEGER PRIMARY KEY AUTOINCREMENT,
-       execution_id INTEGER,
-       kind TEXT NOT NULL,
-       content TEXT NOT NULL,
-       domains TEXT NOT NULL DEFAULT '[]',
-       occurred_at INTEGER NOT NULL
-     );
-     CREATE INDEX IF NOT EXISTS reflections_by_kind ON reflections(kind);";
-
-/// `tasks` DDL at [`SCHEMA_VERSION`] — the latest task-board snapshot per
-/// session, one row per (session, task id), mirrored from the protocol's
-/// `TaskUpdate` snapshots by [`Store::record_task_board`]. UNIQUE
-/// (session_id, task_id) is the upsert key: each snapshot REPLACES a task's
-/// row (board state, not history — the `events` stream already keeps every
-/// snapshot). NOTE: SQL NULLs are pairwise distinct, so rows recorded
-/// without a session id never conflict — dedup only holds per session.
-/// `status`/`owner` carry the protocol's serde snake_case strings (e.g.
-/// `"in_progress"`); `task_id` is the board's per-session ordinal id
-/// ("1", "2", …), read back in `CAST(task_id AS INTEGER)` order.
-const TASKS_DDL: &str = "CREATE TABLE IF NOT EXISTS tasks (
-       id INTEGER PRIMARY KEY AUTOINCREMENT,
-       execution_id INTEGER NOT NULL,
-       session_id TEXT,
-       task_id TEXT NOT NULL,
-       subject TEXT NOT NULL,
-       description TEXT,
-       status TEXT NOT NULL,
-       owner TEXT,
-       updated_at INTEGER NOT NULL,
-       UNIQUE(session_id, task_id)
-     );";
-
-/// `pull_requests` DDL at [`SCHEMA_VERSION`] — one row per tracked pull
-/// request, keyed by URL (the one stable identity across forks/renames).
-/// UNIQUE (url) is the upsert key for [`Store::upsert_pull_request`]: a
-/// later observation of the same PR updates its status/CI verdict in place.
-/// `session_id` is the producing session's registry id, NULL when unknown;
-/// `updated_at` is epoch millis of the latest observation.
-const PULL_REQUESTS_DDL: &str = "CREATE TABLE IF NOT EXISTS pull_requests (
-       id INTEGER PRIMARY KEY AUTOINCREMENT,
-       session_id TEXT,
-       url TEXT NOT NULL,
-       number INTEGER,
-       status TEXT NOT NULL,
-       ci_status TEXT,
-       updated_at INTEGER NOT NULL,
-       UNIQUE(url)
-     );";
-
-/// The full latest schema, applied in one shot to fresh databases only.
-/// Existing files never see this — [`MIGRATIONS`] upgrades them shape by
-/// shape, so this can always describe the CURRENT shape.
-fn create_latest_schema(tx: &rusqlite::Transaction<'_>) -> Result<()> {
-    tx.execute_batch(EXECUTIONS_DDL)?;
-    tx.execute_batch(UNCHANGED_TABLES)?;
-    tx.execute_batch(&events_ddl("events"))?;
-    tx.execute_batch(&telemetry_ddl("telemetry"))?;
-    tx.execute_batch(&files_touched_ddl("files_touched"))?;
-    tx.execute_batch(RULES_TABLE)?;
-    tx.execute_batch(TELEMETRY_INDEX)?;
-    tx.execute_batch(MEMORY_CITATIONS_DDL)?;
-    tx.execute_batch(AGENT_USES_DDL)?;
-    tx.execute_batch(SKILL_USAGE_DDL)?;
-    tx.execute_batch(MCP_USAGE_DDL)?;
-    tx.execute_batch(TOOL_CALLS_DDL)?;
-    tx.execute_batch(EXECUTION_REFLECTION_DDL)?;
-    tx.execute_batch(REFLECTIONS_DDL)?;
-    tx.execute_batch(TASKS_DDL)?;
-    tx.execute_batch(PULL_REQUESTS_DDL)?;
-    Ok(())
-}
-
-fn table_exists(conn: &Connection, table: &str) -> Result<bool> {
-    let count: i64 = conn.query_row(
-        "SELECT count(*) FROM sqlite_master WHERE type = 'table' AND name = ?",
-        params![table],
-        |row| row.get(0),
-    )?;
-    Ok(count > 0)
-}
-
-/// Whether any store-owned table exists — distinguishes a fresh empty file
-/// (create latest schema directly) from a legacy pre-versioning file (run
-/// the migration list), since both carry `user_version` 0.
-fn any_store_table_exists(conn: &Connection) -> Result<bool> {
-    let placeholders = TABLES.iter().map(|_| "?").collect::<Vec<_>>().join(", ");
-    let count: i64 = conn.query_row(
-        &format!(
-            "SELECT count(*) FROM sqlite_master WHERE type = 'table' AND name IN ({placeholders})"
-        ),
-        rusqlite::params_from_iter(TABLES),
-        |row| row.get(0),
-    )?;
-    Ok(count > 0)
-}
-
-fn column_exists(conn: &Connection, table: &str, column: &str) -> Result<bool> {
-    let count: i64 = conn.query_row(
-        "SELECT count(*) FROM pragma_table_info(?) WHERE name = ?",
-        params![table, column],
-        |row| row.get(0),
-    )?;
-    Ok(count > 0)
-}
-
-/// v0 → v1: retrofit the UNIQUE constraints the write paths have always
-/// assumed (see [`events_ddl`]/[`telemetry_ddl`]), deduping first — a
-/// constraint cannot land on a table holding historic duplicates.
-///
-/// Keep-rule: the newest row per natural key — `max(rowid)`, which is
-/// insertion order. A duplicate key can only come from a double-write of
-/// the same logical record (the writers' counters are monotonic per
-/// execution), and readers want the writer's final word: replay renders one
-/// event per stream position, and analytics prices one row per committed
-/// call — exactly the row an upsert would have retained.
-///
-/// SQLite cannot ALTER a UNIQUE constraint in, so both tables are rebuilt
-/// per the documented procedure (lang_altertable §7): create-new →
-/// INSERT SELECT → DROP old → RENAME. The old tables' indexes drop with
-/// them; `telemetry_by_model` is recreated and `events_by_execution` is
-/// superseded by the UNIQUE constraint's implicit index on exactly its
-/// columns. No store table declares foreign keys in either direction, so
-/// the rebuild moves no FK edges — but the runner still follows the full §7
-/// procedure (`foreign_keys` OFF outside the transaction, `foreign_key_check`
-/// before commit) so a future FK-bearing schema cannot be corrupted by this
-/// path.
-///
-/// A v0 file is not guaranteed to hold every table (partial files exist —
-/// e.g. pre-drift fixtures with only `telemetry`), so missing tables are
-/// created fresh in the v1 shape: empty, nothing to dedupe.
-fn migrate_v0_to_v1(tx: &rusqlite::Transaction<'_>) -> Result<()> {
-    tx.execute_batch(UNCHANGED_TABLES)?;
-    // executions changed shape in v8 (the session_id column), so it left
-    // UNCHANGED_TABLES — but a v1 database has its ERA's shape, which this
-    // step must keep producing (the v8 ALTER later in the chain runs
-    // against it).
-    tx.execute_batch(
-        "CREATE TABLE IF NOT EXISTS executions (
-           id INTEGER PRIMARY KEY AUTOINCREMENT,
-           kind TEXT NOT NULL,
-           prompt TEXT NOT NULL,
-           provider TEXT NOT NULL,
-           model TEXT NOT NULL,
-           started_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP,
-           finished_at TEXT,
-           outcome TEXT,
-           cost_usd REAL NOT NULL DEFAULT 0
-         );",
-    )?;
-    // files_touched changed shape again in v2, so it left UNCHANGED_TABLES —
-    // but a v1 database has its ERA's shape, which this step must keep
-    // producing (the v2 rebuild right after runs against it).
-    tx.execute_batch(
-        "CREATE TABLE IF NOT EXISTS files_touched (
-           execution_id INTEGER NOT NULL,
-           path TEXT NOT NULL,
-           ops TEXT NOT NULL
-         );",
-    )?;
-    // New executions must never reuse an id that historic rows already
-    // reference: a reused id mis-attributes those orphaned rows to the new
-    // run, and — with the UNIQUE keys this migration retrofits — collides
-    // with their (execution_id, seq/step) positions. A partial v0 file can
-    // hold events/telemetry that outlive their executions table (whose
-    // AUTOINCREMENT counter then restarts at 1), so the counter is seeded
-    // past every execution id in sight. sqlite_sequence exists here:
-    // creating any AUTOINCREMENT table (executions, just ensured) creates
-    // it, and its content is plain-DML-writable by design.
-    let max_in_executions: Option<i64> =
-        tx.query_row("SELECT max(id) FROM executions", [], |row| row.get(0))?;
-    let mut max_execution_id = max_in_executions.unwrap_or(0);
-    // events and telemetry may still be missing here (they are ensured or
-    // rebuilt below), so each referencing table is probed individually.
-    for table in ["events", "telemetry", "files_touched"] {
-        if table_exists(tx, table)? {
-            let max_id: Option<i64> = tx.query_row(
-                &format!("SELECT max(execution_id) FROM {table}"),
-                [],
-                |row| row.get(0),
-            )?;
-            max_execution_id = max_execution_id.max(max_id.unwrap_or(0));
-        }
-    }
-    if max_execution_id > 0 {
-        let seeded = tx.execute(
-            "UPDATE sqlite_sequence SET seq = ?1 WHERE name = 'executions' AND seq < ?1",
-            params![max_execution_id],
-        )?;
-        if seeded == 0 {
-            // No row updated: either the counter is already past the ids
-            // (nothing to do) or the counter row does not exist yet.
-            let exists: i64 = tx.query_row(
-                "SELECT count(*) FROM sqlite_sequence WHERE name = 'executions'",
-                [],
-                |row| row.get(0),
-            )?;
-            if exists == 0 {
-                tx.execute(
-                    "INSERT INTO sqlite_sequence (name, seq) VALUES ('executions', ?1)",
-                    params![max_execution_id],
-                )?;
-            }
-        }
-    }
-    if table_exists(tx, "events")? {
-        tx.execute_batch(&events_ddl("events_v1"))?;
-        tx.execute_batch(
-            "INSERT INTO events_v1 (execution_id, seq, ts, event_type, payload)
-             SELECT execution_id, seq, ts, event_type, payload FROM events
-             WHERE rowid IN (SELECT max(rowid) FROM events GROUP BY execution_id, seq);
-             DROP TABLE events;
-             ALTER TABLE events_v1 RENAME TO events;",
-        )?;
-    } else {
-        tx.execute_batch(&events_ddl("events"))?;
-    }
-    if table_exists(tx, "telemetry")? {
-        // Pre-drift files lack estimated_input_tokens; the rebuild
-        // backfills 0 = "no estimate was taken", which drift_samples
-        // excludes as signal-free — same semantics the old ALTER-based
-        // migration gave those rows.
-        let estimated = if column_exists(tx, "telemetry", "estimated_input_tokens")? {
-            "estimated_input_tokens"
-        } else {
-            "0"
-        };
-        tx.execute_batch(&telemetry_ddl("telemetry_v1"))?;
-        tx.execute_batch(&format!(
-            "INSERT INTO telemetry_v1 (execution_id, step, ts, provider, model, input_tokens,
-               estimated_input_tokens, output_tokens, cache_read_tokens, cache_miss_tokens,
-               cache_write_tokens, cost_usd, duration_ms, retries, tool_calls)
-             SELECT execution_id, step, ts, provider, model, input_tokens,
-               {estimated}, output_tokens, cache_read_tokens, cache_miss_tokens,
-               cache_write_tokens, cost_usd, duration_ms, retries, tool_calls
-             FROM telemetry
-             WHERE rowid IN (SELECT max(rowid) FROM telemetry GROUP BY execution_id, step);
-             DROP TABLE telemetry;
-             ALTER TABLE telemetry_v1 RENAME TO telemetry;",
-        ))?;
-    } else {
-        tx.execute_batch(&telemetry_ddl("telemetry"))?;
-    }
-    tx.execute_batch(TELEMETRY_INDEX)?;
-    Ok(())
-}
-
-/// v1 → v2: `files_touched` grows per-file line-delta totals and the ordered
-/// JSON audit log ([`FileTouchRow`]), plus the UNIQUE (execution_id, path)
-/// key its writer has always assumed (the ledger emits exactly one record
-/// per normalized path per execution). SQLite cannot ALTER a UNIQUE
-/// constraint in, so the table is rebuilt per lang_altertable §7 with the
-/// same newest-row keep-rule as [`migrate_v0_to_v1`]. Legacy rows predate
-/// line telemetry and are backfilled with the column defaults: zero deltas,
-/// `'[]'` audit log.
-fn migrate_v1_to_v2(tx: &rusqlite::Transaction<'_>) -> Result<()> {
-    if table_exists(tx, "files_touched")? {
-        tx.execute_batch(&files_touched_ddl("files_touched_v2"))?;
-        tx.execute_batch(
-            "INSERT INTO files_touched_v2 (execution_id, path, ops)
-             SELECT execution_id, path, ops FROM files_touched
-             WHERE rowid IN (SELECT max(rowid) FROM files_touched GROUP BY execution_id, path);
-             DROP TABLE files_touched;
-             ALTER TABLE files_touched_v2 RENAME TO files_touched;",
-        )?;
-    } else {
-        // Partial v1 files exist just like partial v0 files: nothing to
-        // rebuild, create the v2 shape fresh.
-        tx.execute_batch(&files_touched_ddl("files_touched"))?;
-    }
-    Ok(())
-}
-
-/// v2 → v3: the `memory_citations` table. Purely additive — no existing
-/// table is touched, so no rebuild, no dedupe, no backfill.
-fn migrate_v2_to_v3(tx: &rusqlite::Transaction<'_>) -> Result<()> {
-    tx.execute_batch(MEMORY_CITATIONS_DDL)?;
-    tx.execute_batch(RULES_TABLE)?;
-    Ok(())
-}
-
-/// v3 → v4: the `agent_uses` invocation log (agent-version usage telemetry,
-/// drained per execution like `files_touched`). Purely additive — no
-/// existing table changes shape, so no §7 rebuild is needed; `IF NOT
-/// EXISTS` also covers a partial file that somehow already grew the table.
-fn migrate_v3_to_v4(tx: &rusqlite::Transaction<'_>) -> Result<()> {
-    tx.execute_batch(AGENT_USES_DDL)?;
-    Ok(())
-}
-
-/// v4 → v5: the additive `skill_usage` table (skill-version usage telemetry,
-/// SKILLS tab). Purely additive, mirroring [`migrate_v3_to_v4`].
-fn migrate_v4_to_v5(tx: &rusqlite::Transaction<'_>) -> Result<()> {
-    tx.execute_batch(SKILL_USAGE_DDL)?;
-    Ok(())
-}
-
-/// v5 → v6: the `mcp_usage` table. Purely additive — no existing table is
-/// touched, so no rebuild, no dedupe, no backfill.
-fn migrate_v5_to_v6(tx: &rusqlite::Transaction<'_>) -> Result<()> {
-    tx.execute_batch(MCP_USAGE_DDL)?;
-    Ok(())
-}
-
-/// v6 → v7: the additive data-plane tables — `tool_calls`,
-/// `execution_reflection`, and `reflections`. No existing table changes shape.
-fn migrate_v6_to_v7(tx: &rusqlite::Transaction<'_>) -> Result<()> {
-    tx.execute_batch(TOOL_CALLS_DDL)?;
-    tx.execute_batch(EXECUTION_REFLECTION_DDL)?;
-    tx.execute_batch(REFLECTIONS_DDL)?;
-    Ok(())
-}
-
-/// v7 → v8: the session plane. `executions` grows the nullable `session_id`
-/// column linking each per-turn row to the cross-process session registry id
-/// (a plain ADD COLUMN — nullable, no default rewrite, so no §7 rebuild),
-/// plus the `executions_by_session` index and the additive `tasks` and
-/// `pull_requests` tables. The ALTER is guarded by a column probe, matching
-/// the house `IF NOT EXISTS` tolerance for partial files that somehow
-/// already grew the shape.
-fn migrate_v7_to_v8(tx: &rusqlite::Transaction<'_>) -> Result<()> {
-    if !column_exists(tx, "executions", "session_id")? {
-        tx.execute_batch("ALTER TABLE executions ADD COLUMN session_id TEXT;")?;
-    }
-    tx.execute_batch(
-        "CREATE INDEX IF NOT EXISTS executions_by_session
-           ON executions(session_id, id);",
-    )?;
-    tx.execute_batch(TASKS_DDL)?;
-    tx.execute_batch(PULL_REQUESTS_DDL)?;
-    Ok(())
-}
-
-/// Run one migration in its own transaction, stamping `user_version` before
-/// commit so version and shape can never disagree on disk. The caller has
-/// already suspended foreign-key enforcement (a no-op inside a
-/// transaction).
-fn apply_migration(conn: &mut Connection, migration: Migration, target: i64) -> Result<()> {
-    let tx = conn.transaction()?;
-    migration(&tx)?;
-    // lang_altertable §7 requires a full FK audit before committing work
-    // done with enforcement off. No store table declares foreign keys
-    // today, so this passes trivially — it is what keeps the runner safe
-    // for schemas that will.
-    let violations: i64 =
-        tx.query_row("SELECT count(*) FROM pragma_foreign_key_check", [], |row| {
-            row.get(0)
-        })?;
-    if violations > 0 {
-        return Err(StoreError(format!(
-            "migration to schema version {target} would leave {violations} \
-             foreign-key violation(s); rolling back"
-        )));
-    }
-    tx.pragma_update(None, "user_version", target)?;
-    tx.commit().map_err(StoreError::from)
 }
 
 pub struct Store {
